requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
av==12.3.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
csvw==3.5.1
ctranslate2==4.4.0
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
dlinfo==2.0.0
espeakng-loader==0.2.4
executing==2.2.1
faster-whisper==1.0.2
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.53.1
fsspec==2025.9.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
ipykernel==6.30.1
ipython==8.27.0
isodate==0.7.2
jedi==0.19.2
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kokoro-onnx==0.4.9
language-tags==1.2.0
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.1.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.106.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
phonemizer-fork==3.3.2
pillow==10.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
PyYAML==6.0.2
pyzmq==27.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rfc3986==1.5.0
rpds-py==0.27.1
rsa==4.9.1
rubicon-objc==0.5.2
segments==2.3.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
